pyproject.toml
[tool.poetry]
name = "jamii"
version = "0.1.0"
description = "A financial group management API"
authors = ["Your Name <dankhareem@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
sqlalchemy = "^2.0.35"
pydantic = {extras = ["email"], version = "^2.9.2"}
databases = {extras = ["postgresql"], version = "^0.9.0"}
psycopg2-binary = "^2.9.9"
alembic = "^1.13.3"
loguru = "^0.7.2"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.5.2"
python-multipart = "^0.0.11"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.30.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
